requirements.txt
langchain-openai
langchain_google_genai
langchain
python-dotenv
notebook 
jupyterlab
ipykernel
notebook
openai
langchain-core
langchain-community
langchainhub
wikipedia
langchain_tavily
langchain_groq
arxiv
